Add action input: post_pr_comment
This option is `true` by default, but if it's set to `false` it will
disable comments being posted on pull requests after a successful deploy.

This gives users more control, for example if they wish to implement
a step in their GitHub Actions workflow which posts a pull request comment.